feat: enhance component organization and simplify imports; add MobileSidebar component
