+++ src/components/Header.tsx
@@ -1,10 +1,10 @@
 import { component$, useSignal, $ } from "@builder.io/qwik";
-import { Link, useLocation, useNavigate } from "@builder.io/qwik-city";
+import { Link, useLocation } from "@builder.io/qwik-city";
 import { clsx } from "clsx";
+import { MobileSidebar } from "./MobileSidebar";
 
 export const Header = component$(() => {
   const loc = useLocation();
-  const nav = useNavigate();
   const isMobileMenuOpen = useSignal(false);
 
   const isActive = (path: string) => {
@@ -19,15 +19,6 @@ export const Header = component$(() => {
 
   const toggleMobileMenu = $(() => {
     isMobileMenuOpen.value = !isMobileMenuOpen.value;
-  });
-
-  const closeMobileMenu = $(() => {
-    isMobileMenuOpen.value = false;
-  });
-
-  const handleMobileNavigation = $((path: string) => {
-    isMobileMenuOpen.value = false;
-    nav(path);
   });
 
   return (
@@ -107,80 +98,7 @@ export const Header = component$(() => {
         </nav>
       </header>
 
-      {/* Mobile Menu Overlay */}
-      {isMobileMenuOpen.value && (
-        <>
-          {/* Backdrop */}
-          <div
-            class="fixed inset-0 bg-black bg-opacity-50 z-40 sm:hidden"
-            onClick$={closeMobileMenu}
-          />
-
-          {/* Sidebar */}
-          <div class="fixed top-0 right-0 h-full w-64 bg-white shadow-lg z-50 sm:hidden">
-            <div class="flex items-center justify-between p-4 border-b border-gray-200">
-              <h2 class="text-lg font-semibold text-gray-900">Menu</h2>
-              <button
-                onClick$={closeMobileMenu}
-                class="p-2 rounded-md text-gray-500 hover:text-gray-700 hover:bg-gray-100"
-                aria-label="Close mobile menu"
-              >
-                <svg
-                  class="h-6 w-6"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    stroke-width={2}
-                    d="M6 18L18 6M6 6l12 12"
-                  />
-                </svg>
-              </button>
-            </div>
-
-            <nav class="p-4">
-              <div class="space-y-4">
-                <button
-                  onClick$={() => handleMobileNavigation("/")}
-                  class={clsx(
-                    "block w-full text-left px-3 py-2 rounded-md text-base font-medium transition-colors",
-                    isActive("/")
-                      ? "bg-primary-100 text-primary-900"
-                      : "text-gray-700 hover:bg-gray-100 hover:text-gray-900",
-                  )}
-                >
-                  Home
-                </button>
-                <button
-                  onClick$={() => handleMobileNavigation("/overview")}
-                  class={clsx(
-                    "block w-full text-left px-3 py-2 rounded-md text-base font-medium transition-colors",
-                    isActive("/overview")
-                      ? "bg-primary-100 text-primary-900"
-                      : "text-gray-700 hover:bg-gray-100 hover:text-gray-900",
-                  )}
-                >
-                  Overview
-                </button>
-                <button
-                  onClick$={() => handleMobileNavigation("/articles")}
-                  class={clsx(
-                    "block w-full text-left px-3 py-2 rounded-md text-base font-medium transition-colors",
-                    isActive("/articles")
-                      ? "bg-primary-100 text-primary-900"
-                      : "text-gray-700 hover:bg-gray-100 hover:text-gray-900",
-                  )}
-                >
-                  Articles
-                </button>
-              </div>
-            </nav>
-          </div>
-        </>
-      )}
+      <MobileSidebar isOpen={isMobileMenuOpen} currentPath={loc.url.pathname} />
     </>
   );
 });
