+++ src/components/index.ts
@@ -1,0 +1,7 @@
+// Central export point for all components
+// This file simplifies imports and provides a clean interface
+
+export * from "./Header";
+export * from "./Footer";
+export * from "./MobileSidebar";
+export * from "./router-head/router-head";
